Refactor Allocation algorithms
