+++ components/org.wso2.carbon.sp.jobmanager.core/src/gen/java/org/wso2/carbon/sp/jobmanager/core/model/WorkerMetrics.java
@@ -1,6 +1,6 @@
 /*
  *
- *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *  Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
